+++ src/main/java/com/internship/dto/SalaryDto.java
@@ -2,12 +2,14 @@ package com.internship.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
 public class SalaryDto {
-    float grossSalary;
-    float netSalary;
+    BigDecimal grossSalary;
+    BigDecimal netSalary;
 }
