+++ src/main/java/com/internship/entity/Employee.java
@@ -7,6 +7,7 @@ import com.internship.enums.Gender;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -52,7 +53,7 @@ public class Employee {
     private Gender gender;
 
     @Column(name = "GROSS_SALARY", nullable = false)
-    private float grossSalary;
+    private BigDecimal grossSalary;
 
     @ManyToOne
     @JoinColumn(name = "DEPARTMENT_ID", nullable = false)
